feat(private-web): declutter top-level nav — restore UI into groups, healthchecks into settings
Restore replicas are per-group, so their UI moves out of the top-level
'Restore' nav item and into each group's backup page: a RestoreReplicasSection
(declarations + that group's restore-health, create dialog with the group
fixed) renders under the ready config. The fleet-wide consumer roster +
capabilities move to a Settings > 'Restore consumers' tab. The standalone
/restore-replicas page is removed.

Healthchecks likewise moves from a top-level nav item into Settings >
'Healthchecks' (list + detail), with its internal links repointed.

Backend: /restore_replicas/checks is now per-group (list_recent_for_group);
the unused global /restore_replicas/list handler is dropped. Regenerated
openapi + api-types. e2e updated to drive the new locations.
